vuer_sim_example/sim2real/requirements.txt
mujoco
pyyaml
scipy
onnxruntime
pynput
sshkeyboard
termcolor
loguru
loop_rate_limiters
pygame
meshcat
